@@ -5,7 +5,8 @@ import { EffectsModule } from '@ngrx/effects';
 import { ReadRoutingModule } from '@read/read-routing.module';
 import { ReadComponent } from '@read/read/read.component';
 import { ReadEffects } from '@read/store/read.effects';
-import { READ_STATE, reducer } from '@read/store/read.reducers';
+import { READ_STATE, readReducer } from '@read/store/read.reducers';
+import { PARAGRAPHS_STATE, paragraphsReducer } from '@read/store/paragraphs.reducers';
 import { MaterialModule } from '@shared/material.module';
 
 @NgModule({
@@ -15,7 +16,8 @@ import { MaterialModule } from '@shared/material.module';
     ReadRoutingModule,
     EffectsModule.forFeature([ReadEffects]),
     MaterialModule,
-    StoreModule.forFeature(READ_STATE, reducer),
+    StoreModule.forFeature(READ_STATE, readReducer),
+    StoreModule.forFeature(PARAGRAPHS_STATE, paragraphsReducer),
   ],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
