@@ -5,19 +5,19 @@ import { BookContent } from './read.entities';
 
 export const READ_STATE = 'read';
 
-export interface State {
+export interface ReadState {
   bookContentLoading: boolean;
   bookContent: BookContent;
   errorMessage: string;
 }
 
-export const initialState: State = {
+export const initialState: ReadState = {
   bookContentLoading: false,
   bookContent: null,
   errorMessage: null,
 };
 
-const readReducer = createReducer(
+const reducer = createReducer(
   initialState,
   on(ReadActions.getBook, (state) => ({ ...state, bookContentLoading: true })),
   on(ReadActions.reederServiceGetBook, (state) => ({ ...state })),
@@ -33,6 +33,6 @@ const readReducer = createReducer(
   })),
 );
 
-export function reducer(state: State | undefined, action: Action) {
-  return readReducer(state, action);
+export function readReducer(state: ReadState | undefined, action: Action) {
+  return reducer(state, action);
 }
