@@ -3,9 +3,9 @@ import { createFeatureSelector, createSelector, select } from '@ngrx/store';
 import { pipe } from 'rxjs';
 import { filter } from 'rxjs/operators';
 
-import { READ_STATE, State } from '@read/store/read.reducers';
+import { READ_STATE, ReadState } from '@read/store/read.reducers';
 
-export const selectRead = createFeatureSelector<State>(READ_STATE);
+export const selectRead = createFeatureSelector<ReadState>(READ_STATE);
 
 export const selectBookContent = createSelector(selectRead, (state) => state.bookContent);
 export const selectBookContentLoading = createSelector(selectRead, (state) => state.bookContentLoading);
