@@ -6,7 +6,7 @@ describe('NavService', () => {
   beforeEach(() => TestBed.configureTestingModule({}));
 
   it('should be created', () => {
-    const service: NavService = TestBed.get(NavService);
+    const service: NavService = TestBed.inject(NavService);
     expect(service).toBeTruthy();
   });
 });
